implemented token checking & bug fixing address_typ query and random events_dapp numbers
implemented a function in the ethereumnode.py to check if address is a token

fixed a missing step in the address typ query, where "from"-addresses got not included. this preprocessing step will now done at once for both coloumns.
using set of "to"/"address" and "from" coloumn

fix bug where random decimal numbers are in events_dapp like in chickenHunt in the "to" coloumn